fix: Use longValue for git branch poll interval  - Persist values from the Startup UI: In SettingsGlobalPanel.applySettings(), call MainProject.setGitBranchPollerEnabled(checkbox.isSelected()) and MainProject.setGitBranchPollIntervalMs(((Number) spinner.getValue()).longValue()). Add debug logs already present in setters will record changes.

+++ app/src/main/java/io/github/jbellis/brokk/gui/dialogs/SettingsGlobalPanel.java
@@ -1274,7 +1274,7 @@ public class SettingsGlobalPanel extends JPanel implements ThemeAware, SettingsC
             MainProject.setGitBranchPollerEnabled(branchPollerEnabled);
             logger.debug("Applied Git branch poller enabled: {}", branchPollerEnabled);
         }
-        int intervalValue = ((Number) gitBranchPollerIntervalSpinner.getValue()).intValue();
+        long intervalValue = ((Number) gitBranchPollerIntervalSpinner.getValue()).longValue();
         long prevInterval = MainProject.getGitBranchPollIntervalMs();
         if (intervalValue != prevInterval) {
             MainProject.setGitBranchPollIntervalMs(intervalValue);
